Fix [Clients]: paging thing that was broken Feat [Orders]: client state on orders now works as intended

+++ app/src/main/java/com/poli/prevendasomie/data/repository/ClientsRepositoryImpl.kt
@@ -28,7 +28,7 @@ class ClientsRepositoryImpl
         return local.getNonPaginatedClients()
     }
 
-    override suspend fun getSelectedClient(clientId: Int): ClientesCadastroEntity {
+    override suspend fun getSelectedClient(clientId: Long): ClientesCadastroEntity {
 
         return local.getSelectedClient(clientId = clientId)
     }
