+++ app/src/main/java/com/poli/prevendasomie/domain/repository/LocalDataSource.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface LocalDataSource {
 
-    suspend fun getSelectedClient(clientId: Int): ClientesCadastroEntity
+    suspend fun getSelectedClient(clientId: Long): ClientesCadastroEntity
     fun getNonPaginatedClients(): Flow<List<ClientesCadastroEntity>>
 
     suspend fun insertSelectedCliente(selectedClient: ClientesCadastroEntity)
