+++ app/src/main/java/com/poli/prevendasomie/domain/usecase/clients/GetClientDetailsUseCase.kt
@@ -4,5 +4,5 @@ import com.poli.prevendasomie.domain.model.clientes.ClientesCadastro
 
 interface GetClientDetailsUseCase {
 
-    suspend operator fun invoke(clientId: Int): ClientesCadastro
+    suspend operator fun invoke(clientId: Long): ClientesCadastro
 }
