+++ app/src/main/java/com/poli/prevendasomie/presentation/clientes/client_detail/ClientDetailViewModel.kt
@@ -30,7 +30,7 @@ class ClientDetailViewModel @Inject constructor(
     init {
         viewModelScope.launch(Dispatchers.IO) {
 
-            val clientId = savedStateHandle.get<Int>("clientId")
+            val clientId = savedStateHandle.get<Long>("clientId")
             Log.d("VAL clientId", "$clientId")
             _selectClient.value = clientId?.let {
                 useCase.getSelectedClientUseCase(clientId = clientId)
